Move abi_utils to api_c/utils_c
