+++ libvcx/src/api_c/issuer_credential.rs
@@ -4,8 +4,8 @@ use indy_sys::CommandHandle;
 use libc::c_char;
 
 use crate::{connection, credential_def, issuer_credential, settings};
-use crate::abi_utils::cstring::CStringUtils;
-use crate::abi_utils::runtime::execute;
+use crate::api_c::utils_c::cstring::CStringUtils;
+use crate::api_c::utils_c::runtime::execute;
 use crate::error::prelude::*;
 use crate::utils::error;
 
@@ -849,8 +849,8 @@ pub mod tests {
 
     use agency_client::mocking::AgencyMockDecrypted;
 
-    use crate::abi_utils::return_types_u32;
-    use crate::abi_utils::timeout::TimeoutUtils;
+    use crate::api_c::utils_c::return_types_u32;
+    use crate::api_c::utils_c::timeout::TimeoutUtils;
     use crate::api_c::VcxStateType;
     use crate::settings;
     use crate::utils::constants::*;
