+++ libvcx/src/api_c/mod.rs
@@ -1,9 +1,10 @@
 use std::fmt;
 
+#[macro_use]
+pub mod utils_c;
 pub mod vcx;
 pub mod connection;
 pub mod issuer_credential;
-pub mod utils;
 pub mod proof;
 pub mod credential_def;
 pub mod schema;
@@ -12,6 +13,7 @@ pub mod disclosed_proof;
 pub mod wallet;
 pub mod logger;
 mod filters;
+
 
 /// This macro allows the VcxStateType to be
 /// serialized within serde as an integer (represented as
