+++ libvcx/src/api_c/utils_c/ccallback.rs
