+++ libvcx/src/api_c/utils_c/cstring.rs
