+++ libvcx/src/api_c/utils_c/timeout.rs
