+++ libvcx/src/init.rs
@@ -3,8 +3,7 @@ use indy::ErrorCode;
 use indy::future::Future;
 use indy_sys::WalletHandle;
 
-use crate::{abi_utils, settings, utils};
-use crate::abi_utils::runtime::ThreadpoolConfig;
+use crate::{settings, utils};
 use crate::error::{VcxError, VcxErrorExt, VcxErrorKind, VcxResult};
 use crate::libindy::utils::pool::{create_pool_ledger_config, open_pool_ledger};
 use crate::libindy::utils::wallet::{build_wallet_config, build_wallet_credentials, IssuerConfig, set_wallet_handle, WalletConfig};
@@ -15,13 +14,6 @@ pub struct PoolConfig {
     pub genesis_path: String,
     pub pool_name: Option<String>,
     pub pool_config: Option<String>,
-}
-
-pub fn init_threadpool(config: &str) -> VcxResult<()> {
-    let config: ThreadpoolConfig = serde_json::from_str(config)
-        .map_err(|err| VcxError::from_msg(VcxErrorKind::InvalidJson, format!("Failed to deserialize threadpool config {:?}, err: {:?}", config, err)))?;
-    abi_utils::runtime::init_runtime(config);
-    Ok(())
 }
 
 pub fn enable_vcx_mocks() -> VcxResult<()> {
